cthulhu: crush_node create happy path works

+++ rest-api/calamari_rest/serializers/v2.py
@@ -178,7 +178,7 @@ class CrushRuleSerializer(serializers.Serializer):
 
 class NodeItemSerializer(serializers.Serializer):
     id = serializers.IntegerField()
-    weight = serializers.IntegerField()
+    weight = serializers.FloatField()
     pos = serializers.IntegerField()
 
     class Meta:
